+++ src/index.ts
@@ -4,9 +4,10 @@ import {
     diValue as value,
     diUse as use,
     diFactory as factory,
+    diFunc as func,
 } from "./resolversShorthands";
 import { IDIContainer } from "./types";
 
 export default DIContainer;
 
-export { object, value, use, factory, IDIContainer };
+export { object, value, use, func, factory, IDIContainer };