+++ src/resolversShorthands.ts
@@ -3,7 +3,8 @@ import RawValueResolver from "./resolvers/RawValueResolver";
 import ReferenceResolver from "./resolvers/ReferenceResolver";
 import FactoryResolver, { Factory } from "./resolvers/FactoryResolver";
 import { definitionNameToString } from "./DefinitionName";
-import { ClassOf, ResolverName } from "./types";
+import { ClassOf, ResolverName, WrapWithResolver } from "./types";
+import FunctionResolver from "./resolvers/FunctionResolver";
 
 // shorthands for Definition classes
 
@@ -57,4 +58,19 @@ export function diUse<Custom = void, Name extends ResolverName = ResolverName>(
  */
 export function diFactory<T extends Factory>(factory: T) {
     return new FactoryResolver(factory);
+}
+
+/**
+ * FunctionResolver - allows to use custom function with specified parameters, where parameters are references to
+ * existing dependencies
+ * @param func
+ * @param parameters
+ */
+export function diFunc<T extends (...args: any) => any>(
+    func: T,
+    ...parameters: T extends (...args: infer ArgTypes) => any
+        ? WrapWithResolver<ArgTypes>
+        : never
+) {
+    return new FunctionResolver(func, ...parameters);
 }