+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
+# 2.1.0
+
+## Changed
+
+-   Adds function resolver
+
 # 2.0.0
 
 ## Changed
 
-- Introduced more strict type checks
+-   Introduced more strict type checks